refactor: move css component only with all others component
