+++ ui-particles-angular/projects/ui-particles-angular/ng-package.json
@@ -5,6 +5,6 @@
     "entryFile": "src/public-api.ts",
     "styleIncludePaths": ["../../../node_modules"]
   },
-  "assets": ["index.scss", "src/scss/**/*.scss", "LICENSE.txt", "assets/**/*"],
+  "assets": ["index.scss", "src/scss/**/*.scss", "src/lib/**/!(*.component).scss", "LICENSE.txt", "assets/**/*"],
   "allowedNonPeerDependencies": ["@fontsource/fira-mono", "@fontsource/golos-ui", "@fontsource/material-icons"]
 }
